Create layout component for reusability
